% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_matrix_diag.R
\name{zero_matrix_diag}
\alias{zero_matrix_diag}
\title{Create a diagonal structural zero matrix}
\usage{
zero_matrix_diag(state, basename = "state")
}
\arguments{
\item{state}{A character vector, giving the names of the
states.}

\item{basename}{Prefix used to construct dimension
names.  Defaults to \code{"state"}.}
}
\value{
A square matrix with \code{TRUE} on the
diagonal, \code{FALSE} elsewhere, and
dimnames constructed from \code{states}.
}
\description{
Create a structural zero matrix with \code{TRUE}
on the diagonal and \code{FALSE} everywhere else.
This matrix represents situtation where a movement
between any pair of distinct states is permitted,
but a movement from a state to itself is prohibited.
}
\details{
Diagonal structural zero matrices are often
used when modelling migration, since migration statistics
often show movements between regions, but not movements
within regions.
}
\examples{
states <- c("Thailand", "Malaysia", "Singapore")
zero_matrix_diag(states)
zero_matrix_diag(states, basename = "country")
}
\seealso{
\code{zero_matrix_trans} deals with the
general case.
}
