% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_matrix_nest.R
\name{zero_matrix_nest}
\alias{zero_matrix_nest}
\title{Create a structural zero matrix representing nested
categories}
\usage{
zero_matrix_nest(nest)
}
\arguments{
\item{nest}{A data.frame providing a mapping between
the higher-level and lower-level categories.}
}
\value{
A matrix of \code{TRUE}s and \code{FALSE}s,
with dimnames constructed from \code{nest}.
}
\description{
Create a structural zero matrix representing a situation where
one set of categories is nested within another - eg
counties within states, or detailed occupations within
broad occupational categories.
}
\details{
Nesting can be contrasted with crossing, where each
set of categories are independent of the others,
eg age vs sex.

\code{nest} is a data.frame providing a mapping between
the higher-level categories and lower-level categories.
\code{nest} has two columns: one with the higher-level
categories, and one with the lower-level categories.
Each row gives the higher-level category associated with a
lower-level category. The order of the columns and rows
does not matter.

Each lower-level category can belong to only one
higher-level category.

\code{zero_matrix_nest} returns a logical matrix
indicating whether each lower-level category
nests within the corresponding higher-level category.
}
\examples{
nest <- data.frame(region = c("Asia", "Africa", "Africa"),
                   country = c("India", "Nigeria", "Zimbabwe"))
zero_matrix_nest(nest)

nest <- data.frame(level2 = c("Bacterial infections",
                              "Viral infections",
                              "Infections caused by fungi",
                              "Lip, oral cavity, pharynx",
                              "Digestive organs"),
                   level1 = c("Infectious and parasitic diseases",
                              "Infectious and parasitic diseases",
                              "Infectious and parasitic diseases",
                              "Neoplasms",
                              "Neoplasms"))
zero_matrix_nest(nest)
}
\seealso{
Function \code{\link{add_dim_nest}} adds a
dimension representing the higher-level categories to
an array containing the lower-level categories.
}
